+++ scripts/macros.js
@@ -285,6 +285,7 @@ export {avatarOfBattle} from './macros/2024/classFeatures/cleric/war/avatarOfBat
 export {favoredEnemy} from './macros/2024/classFeatures/ranger/favoredEnemy.js';
 export {tireless} from './macros/2024/classFeatures/ranger/tireless.js';
 export {relentlessHunter} from './macros/2024/classFeatures/ranger/relentlessHunter.js';
+export {naturesVeil} from './macros/2024/classFeatures/ranger/naturesVeil.js';
 //Rogue
 export {sneakAttack} from './macros/2024/classFeatures/rogue/sneakAttack.js';
 export {cunningAction} from './macros/2024/classFeatures/rogue/cunningAction.js';