+++ scripts/macros/2024/classFeatures/ranger/naturesVeil.js
@@ -1,0 +1,4 @@
+export let naturesVeil = {
+    name: 'Nature\'s Veil',
+    version: '1.3.78'
+};