+++ scripts/macros/2024/classFeatures/ranger/preciseHunter.js
@@ -1,0 +1,4 @@
+export let preciseHunter = {
+    name: 'Precise Hunter',
+    version: '1.3.78'
+};